testing scale down of sampled image

+++ android/src/main/java/com/reactnativecommunity/cameraroll/ThumbnailCreatorTask.java
@@ -238,14 +238,14 @@ public class ThumbnailCreatorTask extends GuardedAsyncTask<Void, Void> {
         int resultHeight = requestedHeight;
         float scaleRatio = 1;
         if (bitmapWidth < bitmapHeight) {
-            scaleRatio = requestedWidth / bitmapWidth;
+            scaleRatio = ((float) requestedWidth) / bitmapWidth;
             Log.d("RNCameraRoll", "createScaledBitmap scaleRatio: " + requestedWidth + "/" + bitmapWidth + "=" + scaleRatio);
             resultHeight = (int) (bitmapHeight * scaleRatio);
             Log.d("RNCameraRoll", "scaleAndCropBitmap createScaledBitmap passed parameters - new height: " + (requestedHeight * scaleRatio));
         }
         // if height < width, use requestedHeight as reference for scale
         else if (bitmapHeight < bitmapWidth) {
-            scaleRatio = requestedHeight / bitmapHeight;
+            scaleRatio = ((float) requestedHeight) / bitmapHeight;
             Log.d("RNCameraRoll", "createScaledBitmap scaleRatio: " + requestedHeight + "/" + bitmapHeight + "=" + scaleRatio);
             Log.d("RNCameraRoll", "createScaledBitmap scaleRatio: " + scaleRatio);
             resultWidth = (int) (bitmapWidth * scaleRatio);
